Start to refactor timeline overlay block display
