+++ src/routes/timeline/+page.svelte
@@ -1,5 +1,5 @@
 <script>
-    import Timeline from "$lib/components/Timeline.svelte";
+    import Timeline from "$lib/components/timeline/Timeline.svelte";
 </script>
 
 <div class="page">
